backend/requirements.txt
Django==4.2.20
djangorestframework==3.16.0
django-filter==21.1
Pillow==9.5.0
django-cors-headers==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
djoser==2.1.0
python-decouple==3.8
django-colorfield==0.13.0
django-debug-toolbar==5.1.0